Implemented the dcache_flush_all routine
-- Checkes the CLIDR_EL1 and CCSIDR_EL1 register to loop through all cache levels, sets and ways
and flush them using the DC CISW instruction (clean and invalidate by set/way)
-- Also fix a bug in vsprintf; we need integer division. FP division results in some weird values...

+++ lib/vsprintf.c
@@ -95,7 +95,7 @@ char* itoa2(unsigned long long num, char *str, unsigned int base, bool upcase) {
     int i = 0;
     while(num != 0) {
         unsigned long R = _umod(num, base);
-        num = (unsigned long long)((double)num / (double)base);
+        num = num / base;
         tmp[i++] = (upcase) ? upper[R] : lower[R];
     }
